requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.10
djangorestframework==3.7.3
httplib2==0.10.3
idna==2.6
oauth2==1.9.0.post1
oauthlib==2.0.6
psycopg2==2.7.3.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
tweepy==3.5.0
urllib3==1.22